chore(Routers): cria baserouter firebase.
