+++ functions/src/Routers/baseFirebaseFunction.router.ts
@@ -1,0 +1,39 @@
+import { Request, Response } from "firebase-functions";
+import { IDB } from "../interfaces/db"
+
+export class BaseFirebaseFunctionRouter<T> {
+  constructor (private controller: IDB<T>) {}
+
+  async execute (request: Request, response: Response){
+    let obj = {};
+    const { body, params } = request;
+    const id = params[0];
+
+    switch (request.method) {
+      case "DELETE":
+        obj = await this.controller.delete(id);
+        response.json(obj);
+        break;
+      case "PUT":
+        obj = await this.controller.update(id, body);
+        response.json(obj);
+        break;
+      case "POST":
+        obj = await this.controller.save(body);
+        response.json(obj);
+        break;
+      case "GET":
+        if (id === "/" || !id) {
+          response.json(await this.controller.getAll());
+          break;
+        }
+        response.json(await this.controller.getById(id));
+        break;
+      default:
+        response.json({
+          message: "...",
+        });
+        break;
+    }
+  }
+}
